requirements.txt
appdirs==1.4.3
Django==1.11.1
django-extensions==1.7.9
django-filter==1.0.3
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
Markdown==2.6.8
packaging==16.8
psycopg2==2.7.1
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
